requirements.txt
joblib==1.3.2
langchain==0.0.352
langchain_community==0.0.3
langchain_core==0.1.1
langchain_google_genai==0.0.6
more_itertools==10.1.0
openai==0.28.0
pandas==2.0.2
pydantic==2.5.3
python-dotenv==1.0.0
tqdm==4.65.0
